Nov 22, 2023, 6:00 PM

+++ lab13/Task1/Task1.java
@@ -33,7 +33,7 @@ public class Task1 {
             count++;
         }
 
-        double avg = sum / count;
+        double avg = (double) sum / count;
 
         System.out.println("Sum: " + sum);
         System.out.println("Avg: " + avg);
